Implement MediatR for search functionality and Event Handler to Log the request and search values
The commit restructures the search system in our book library by implementing the MediatR library. It introduces a new SearchModelHandler class that deals with search requests, hence decoupling search logic from the Controller. It also modifies the BookLibraryController and Program to support the MediatR changes and adds a logging mechanism for search queries. Lastly, it switches database context handling to a new Repository class.